+++ app/components/weather/Controls.tsx
