+++ app/constants/index.ts
@@ -1,0 +1,10 @@
+export const INITIAL_CITIES = [
+  { name: "Chapecó", state: "SC" },
+  { name: "Passo Fundo", state: "RS" },
+  { name: "Erechim", state: "RS" },
+  { name: "Carazinho", state: "RS" },
+  { name: "Santo Ângelo", state: "RS" },
+  { name: "Lages", state: "SC" },
+  { name: "Ijuí", state: "RS" },
+  { name: "Vacaria", state: "RS" },
+];
